Fix #1115 - part2 - fix tests and add all_tags
Up until this patch, TemplateWrapper.tags was returning all the tags
defined in the catalog. I think this is wrong and tags shoul only
return the defined tags in the current scope.
Hence, I defined a all_tags method that returns the list of tags defined
in the whole catalog.