Fix #1691 - Realize fails with array of Resource References
The following snippet:
realize( File["/tmp/a","/tmp/b"] )

is parsed into:
AST::Function @name=realize @arguments=
    AST::ASTArray @children = [
       AST::ResourceReference @title= AST::ASTArray @children = [
          String(/tmp/a), String(/tmp/b)
       ]
    ]

When evaluated:
ResourceReference gives -> [ File[/tmp/a], File[/tmp/b] ]

which means the function arguments are:
[[File[/tmp/a], File[/tmp/b]]
after evaluating the @arguments ASTArray of AST::Functions.

Then the collector complains that it can't find the resource
because it is not supposed to work on non-flattened resource
array.

The fix is to flatten in the realize function (although it can
be done more generally in the AST::Function evaluation) before
the resources are given to the Collector.

Signed-off-by: Brice Figureau <brice-puppet@daysofwonder.com>