Add an unique index over resources title, type and host
Rationale: this adds better integrity by telling the database layer
that a resource is unique per host for a combination of type and
title.

CAVEAT for MySQL users:
On MySQL it is not possible to have a large index over TEXT column,
so the unique index covers only the first 255 characters of the
resources title column. This might not be enough for all sites.

Signed-off-by: Brice Figureau <brice-puppet@daysofwonder.com>