Close rails database connection after each compilation
Rationale: Rails likes to cache a lot of things to improve
performance of OLTP applications. Puppet is a long
lived server which doesn't really have an OLTP database workload.
The idea is to close the connection to the database after each
compilation of a node to flush all the active cache to release
some memory for other needs.

Warning: it is not proved this is necessary. This patch has been tested
on production with a small load. YMMV.

Signed-off-by: Brice Figureau <brice-puppet@daysofwonder.com>