Add an unmunge capability to type parameters and properties
Unmunge is the reverse of munge.
While munge allows the type to return a different parameter value
or properties should than the one it was created with, unmunge
does the reverse.

It can be used for instance to store a value in a different
representation but still be able to return genuine value to the
outside world.

Signed-off-by: Brice Figureau <brice-puppet@daysofwonder.com>