Enhance authconfig format to support uri paths
This patch introduces a new set of directive to the authconfig
parser/file format:

path /uripath
This directive declares a new kind of ACL based on the uri path.

method save, find
This directive which is to be used under path directive restricts a
path ACL to only some REST verbs.

The ACL path system matches on the longest path prefix possible.
If no path are matching, then the authorization is not allowed.
The same if no ACL matches for the given REST verb.

Signed-off-by: Brice Figureau <brice-puppet@daysofwonder.com>