Fix #2230 - Puppetmaster not collecting exported resources
In 0.25 we convert Puppet::Parser::Resource in Puppet::Resource.
Those are stored in the db with the rails code.
When we convert a Puppet::Parser::Resource to a Puppet::Resource
we didn't copy over the exported attribute, so we were writing or
updating the database with exported=NULL for this resource,
so no host were able to collect it.

Signed-off-by: Brice Figureau <brice-puppet@daysofwonder.com>