Fix #2246 - Array tagged resources can't be collected or exported
I don't know why we imposed the restriction that we shouldn't match
with parameter containing arrays in exported mode.
That doesn't seem right, as the produced rails query works fine with
arrays.

Note: the user tags are not stored in the rails database except under
the special resource parameter tag. This also doesn't seem right.

Signed-off-by: Brice Figureau <brice-puppet@daysofwonder.com>