Fix #2459 - puppetdoc added namespace classes as RDoc modules instead of classes
Puppetdoc while parsing a class name creates intermediate RDoc classes
to form a hierarchy. Due to a coding error, those intermediate classes
were created as RDoc modules.
Later, when puppetdoc was parsing the definition of one of this class,
it wasn't finding it in RDoc list of classes (since it was recorded
as a module).
Puppetdoc was then creating documentation for two objects of the same
name.

Signed-off-by: Brice Figureau <brice-puppet@daysofwonder.com>