Fix #2664 - Mathematic expressions mis-lexed as Regex
This is not the real fix. It is just an hot-fix to limit
the issue.
The issue is that the lexer regexes have precedences over simple
'/' (divide).
In the following expression:
$var = 4096 / 4
$var2 = "/tmp/file"

The / 4... part is mis-lexed as a regex instead of a mathematical
expression.
The current fix limits regex to one-line.

Signed-off-by: Brice Figureau <brice-puppet@daysofwonder.com>