Allows puppet:// content client streaming
When a file source property requests a file content, the request is
routed (indirected) to P::I::FileContent::Rest.
This class instead of relying on the general Rest subsystem to get a
raw response containing the full file content, creates a special
model using a DeferredResponse.
Later when the file type wants to write the file on disk, it "streams"
it through the DeferredResponse.
The DeferredResponse lazily executes the network requests. But due
to the poor Net::HTTP implementation, we're forced to use a new Thread
to do the streaming. Each read chunk is passed in a Thread SizedQueue
which is then fetch by the main tread doing the on disk store.
The new checksum is computed at the same time chunk by chunk.

Signed-off-by: Brice Figureau <brice-puppet@daysofwonder.com>