Fix #4207 - Storeconfigs serialize the env instance instead of name
Environments used to be only some strings, but now we pass around
environment instances which have references to the known resource types
and other information.
When saving the host information in the storeconfigs database, we were
using this environment instance which was yamlized by AR and stored instead
of using the environment name. This could result for storing a potentially
large information in the environment column of the hosts table.

Signed-off-by: Brice Figureau <brice-puppet@daysofwonder.com>