Fix #4458 - Do not dump the whole environment when instances can't be found
When generating the error message when we can't find any instances for a
search request, we were "inspect"ing the request which now contains
an environment instance which itself contains a lots of things (including
all the known resource types).
Thus it was generating a very large "reason" (ie the HTTP error string).
On some environments (ie proxied mongrel) this was too large and the
proxy would produce an error 500.

I just changed the error message to just log the indirection name and
request key (which should be enough to understand what is wrong).

Signed-off-by: Brice Figureau <brice-puppet@daysofwonder.com>