Fix #4339 - Save a last run report summary to $statedir/last_run_summary.yaml
Once a configuration run is done, puppetd will save on the node a
yaml summary report roughly akin to:

---
  time:
    notify: 0.001025
    last_run: 1289561427
    schedule: 0.00071
    config_retrieval: 0.039518
    filebucket: 0.000126
  resources:
    changed: 1
    total: 8
    out_of_sync: 1
  events:
    total: 1
    success: 1
  changes:
    total: 1

This is almost an hash version of the current --summarize output, with
the notable exception that the time section includes the last run unix
timestamp.

The whole idea is to be able to monitor locally if a puppetd does its job.
For instance this could be used in a nagios check or to send an SNMP trap.
The last_run information might help detect staleness, and this summary
can also be used for performance monitoring (ie time section).
The resource section can also show the number of failed resources.

Signed-off-by: Brice Figureau <brice-puppet@daysofwonder.com>