Add resource isolation to puppet device/agent
This way it is not possible for the puppet device application to
modify the current host it is running on, as only resource compatible
with remote network device can be applied (all the other will be skipped).
The only resource type appliable on both host and network device is
"schedule".

Signed-off-by: Brice Figureau <brice-puppet@daysofwonder.com>