Proof of concept auth plugin: HTTP Basic auth
This auth plugin uses HTTP basic authentication to authenticate
nodes to the master.
It doesn't reuse any SSL infrastructure, as such the communication
between the node and the master happens in clear-text.

For webrick masters, the admin must create a $vardir/htpasswd with one
entry per node.
For rack/mongrel. the admin must configure the reverse proxy or apache
to load the same htpasswd.

On the client side, the user must configure --http_basic_password to match
the entry in the htpasswd file.

Both the master and the client needs to be started with the options:
--auth http_basic

Signed-off-by: Brice Figureau <brice-puppet@daysofwonder.com>