Fix #7710 - Allow to set any owner/group in file settings
It is impossible to set an owner/group different than root or service
for any file/directory settings (both from defaults or by speficying
those in the configuration file).
This has been introduced in commit 06fcec to prevent users to
set invalid values.
But for some settings, it might be interesting to use other owners
and groups than root/service.

This patch allows individual settings to remove this restriction by
adding :allow_any_owners_groups boolean property to their defaults.
If this property is false or not present, the default behavior is
used. If this property is true, any combination of owner and groups
is allowed.

Signed-off-by: Brice Figureau <brice-puppet@daysofwonder.com>